# sogapeint-CRM/sogapaint-crm/src/assets/swagger-docs/swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "API Sogapeint",
      "version": "1.0.0",
      "description": "API backend de l'application Sogapeint."
    },
    "host": "localhost:3000",
    "basePath": "/",
    "schemes": [
      "http"
    ],
    "paths": {
...
      "/addUser": {
        "post": {
          "description": "",
          "parameters": [
            {
              "name": "authorization",
              "in": "header",
              "type": "string"
            },
            {
              "name": "body",
              "in": "body",
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "example": "any"
                  },
                  "password": {
                    "example": "any"
                  },
                  "firstname": {
                    "example": "any"
                  },
                  "lastname": {
                    "example": "any"
                  },
                  "phone": {
                    "example": "any"
                  },
                  "company": {
                    "example": "any"
                  },
                  "role": {
                    "example": "any"
                  },
                  "active": {
                    "example": "any"
                  },
                  "authorized_connection": {
                    "example": "any"
                  }
                }
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Created"
            },
            "400": {
              "description": "Bad Request"
            },
            "401": {
              "description": "Unauthorized"
            },
            "403": {
              "description": "Forbidden"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
      "/allUsers": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "authorization",
              "in": "header",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "401": {
              "description": "Unauthorized"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
      "/user/{userId}": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "userId",
              "in": "path",
              "required": true,
              "type": "string"
            },
            {
              "name": "authorization",
              "in": "header",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "401": {
              "description": "Unauthorized"
            },
            "404": {
              "description": "Not Found"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        },
        "put": {
          "description": "",
          "parameters": [
            {
              "name": "userId",
              "in": "path",
              "required": true,
              "type": "string"
            },
            {
              "name": "authorization",
              "in": "header",
              "type": "string"
            },
            {
              "name": "body",
              "in": "body",
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "example": "any"
                  },
                  "firstname": {
                    "example": "any"
                  },
                  "lastname": {
                    "example": "any"
                  },
                  "phone": {
                    "example": "any"
                  },
                  "company": {
                    "example": "any"
                  },
                  "role": {
                    "example": "any"
                  },
                  "active": {
                    "example": "any"
                  },
                  "authorized_connection": {
                    "example": "any"
                  }
                }
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "401": {
              "description": "Unauthorized"
            },
            "403": {
              "description": "Forbidden"
            },
            "404": {
              "description": "Not Found"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        },
        "delete": {
          "description": "",
          "parameters": [
            {
              "name": "userId",
              "in": "path",
              "required": true,
              "type": "string"
            },
            {
              "name": "authorization",
              "in": "header",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "401": {
              "description": "Unauthorized"
            },
            "403": {
              "description": "Forbidden"
            },
            "404": {
              "description": "Not Found"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
      "/user-search": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "authorization",
              "in": "header",
              "type": "string"
            },
            {
              "name": "q",
              "in": "query",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "401": {
              "description": "Unauthorized"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
...
      "/users/addUser": {
        "post": {
          "description": "",
          "parameters": [
            {
              "name": "authorization",
              "in": "header",
              "type": "string"
            },
            {
              "name": "body",
              "in": "body",
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "example": "any"
                  },
                  "password": {
                    "example": "any"
                  },
                  "firstname": {
                    "example": "any"
                  },
                  "lastname": {
                    "example": "any"
                  },
                  "phone": {
                    "example": "any"
                  },
                  "company": {
                    "example": "any"
                  },
                  "role": {
                    "example": "any"
                  },
                  "active": {
                    "example": "any"
                  },
                  "authorized_connection": {
                    "example": "any"
                  }
                }
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Created"
            },
            "400": {
              "description": "Bad Request"
            },
            "401": {
              "description": "Unauthorized"
            },
            "403": {
              "description": "Forbidden"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
      "/users/allUsers": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "authorization",
              "in": "header",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "401": {
              "description": "Unauthorized"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
      "/users/user/{userId}": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "userId",
              "in": "path",
              "required": true,
              "type": "string"
            },
            {
              "name": "authorization",
              "in": "header",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "401": {
              "description": "Unauthorized"
            },
            "404": {
              "description": "Not Found"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        },
        "put": {
          "description": "",
          "parameters": [
            {
              "name": "userId",
              "in": "path",
              "required": true,
              "type": "string"
            },
            {
              "name": "authorization",
              "in": "header",
              "type": "string"
            },
            {
              "name": "body",
              "in": "body",
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "example": "any"
                  },
                  "firstname": {
                    "example": "any"
                  },
                  "lastname": {
                    "example": "any"
                  },
                  "phone": {
                    "example": "any"
                  },
                  "company": {
                    "example": "any"
                  },
                  "role": {
                    "example": "any"
                  },
                  "active": {
                    "example": "any"
                  },
                  "authorized_connection": {
                    "example": "any"
                  }
                }
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "401": {
              "description": "Unauthorized"
            },
            "403": {
              "description": "Forbidden"
            },
            "404": {
              "description": "Not Found"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        },
        "delete": {
          "description": "",
          "parameters": [
            {
              "name": "userId",
              "in": "path",
              "required": true,
              "type": "string"
            },
            {
              "name": "authorization",
              "in": "header",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "401": {
              "description": "Unauthorized"
            },
            "403": {
              "description": "Forbidden"
            },
            "404": {
              "description": "Not Found"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
      "/users/user-search": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "authorization",
              "in": "header",
              "type": "string"
            },
            {
              "name": "q",
              "in": "query",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "401": {
              "description": "Unauthorized"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
...
      "/scrape/company-list/{companyName}": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "companyName",
              "in": "path",
              "required": true,
              "type": "string"
            },
            {
              "name": "region",
              "in": "query",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
      "/scrape/enriched-company/{companyName}/{postalCode}": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "companyName",
              "in": "path",
              "required": true,
              "type": "string"
            },
            {
              "name": "postalCode",
              "in": "path",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "404": {
              "description": "Not Found"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
...
      "/company-list/{companyName}": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "companyName",
              "in": "path",
              "required": true,
              "type": "string"
            },
            {
              "name": "region",
              "in": "query",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      },
      "/enriched-company/{companyName}/{postalCode}": {
        "get": {
          "description": "",
          "parameters": [
            {
              "name": "companyName",
              "in": "path",
              "required": true,
              "type": "string"
            },
            {
              "name": "postalCode",
              "in": "path",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "404": {
              "description": "Not Found"
            },
            "500": {
              "description": "Internal Server Error"
            }
          }
        }
      }
    }
  }